Create a_04_prepPM25dataset
Created a complete dataset for Kamloops and for Kelowna that has hourly PM2.5 concentrations attributed to a WFS designation (WFS or no-WFS) and the corresponding sampler period. 

This was importance because there was a misaligbment in temporal resolution between the HOURLY data, DAILY WFS threshold, and 1 WEEK/2 WEEK chemical sampler periods.